Redesigned how we handle reproducibility of Conda environments
The conda environment, converted into an archive with conda-pack,
will not be stored with git LFS within the repository. The new script
update-environment-archive.sh re-creates environment.tar.gz from
environment.yaml, commits the newly created archive, and stores the
revision of newly committed file in environment.rev. The latter is
necessary for the (to-be-written) script install-environment.sh, which
will download and unpack the revision listed in environment.rev. Such
a script is necessary because GitHub-created repository archives don't
currently include files managed with git LFS.